Перенёс масштаб в editor и выделил в отдельный модуль effects-data
