+++ js/main.js
@@ -1,5 +1,6 @@
 import {createPictures} from './pictures-preview.js';
 import {data} from './random-data.js';
 import './img-upload.js';
+import './editor.js';
 
 createPictures(data);
